+++ frontend/src/components/polls/list/PollNoPolls.tsx
