+++ frontend/src/components/polls/list/PollResults.tsx
