+++ frontend/src/components/polls/list/list_polls.css
