+++ frontend/src/components/polls/table/PollStateIcon.tsx
