+++ frontend/src/components/polls/table/PollsTable.tsx
