+++ frontend/src/components/polls/table/polls-table.css
